5 fix code smell

+++ Services/IOrderService.cs
@@ -4,8 +4,8 @@ namespace LegacyOrderService.Services
 {
     public interface IOrderService
     {
-        Order CreateOrder(string customerName, string productName, int quantity, double price);
-        double CalculateTotal(Order order);
+        Order CreateOrder(string customerName, string productName, int quantity, decimal price);
+        decimal CalculateTotal(Order order);
         void SaveOrder(Order order);
     }
 }
