+++ Services/IProductService.cs
@@ -2,6 +2,6 @@ namespace LegacyOrderService.Services
 {
     public interface IProductService
     {
-        double GetProductPrice(string productName);
+        decimal GetProductPrice(string productName);
     }
 }
